% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegForest.r, R/SurvForest.R
\name{RegForest}
\alias{RegForest}
\alias{SurvForest}
\title{RegForest}
\usage{
RegForest(x, y, ncat, obs.w, var.w, resample.preset, param, ...)

SurvForest(x, y, censor, ncat, obs.w, var.w, resample.preset, param, ...)
}
\description{
Internal function for fitting regression forest

Internal function for fitting regression forest
}
\keyword{internal}
